Properly handle lack of fanotify kernel support in perl tests
As reported by Martin von Wittich in CPAN RT ticket 105259, the module's test
script(s) always failed due to lacking kernel support for (one of the)
fanotify features.

Additionally, cpantesters.org reported issues with oldish perl versions.

With this commit,
a) the test script try to check whether /proc/config.gz or
/boot/config-`uname -r` exist and contain information about configured
fanotify features, and accept respective failure return values accordingly
b) the module depends on perl 5.13 or newer; as fanotify has not been in the
kernel before 2.6.37 -- and that is newer than perl 5.13 -- any system
with a matching kernel should have a newer perl as well.

Thanks to Martin for reporting the issue.